teamroster pkg -> team, split into two files
