+++ team/team.go
@@ -1,14 +1,12 @@
-package teamroster
+package team
 
 import (
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"os"
 	"path/filepath"
 
-	"github.com/BurntSushi/toml"
 	"github.com/fluidkeys/fluidkeys/fingerprint"
 	"github.com/gofrs/uuid"
 )
@@ -96,25 +94,6 @@ func loadTeamRoster(filename string) (*Team, error) {
 	}
 
 	return team, nil
-}
-
-// Parse parses the team roster's TOML data, returning a Team or an error
-func Parse(r io.Reader) (*Team, error) {
-	var parsedTeam Team
-	metadata, err := toml.DecodeReader(r, &parsedTeam)
-
-	if err != nil {
-		return nil, fmt.Errorf("error in toml.DecodeReader: %v", err)
-	}
-
-	if len(metadata.Undecoded()) > 0 {
-		// found config variables that we don't know how to match to
-		// the Team structure
-		return nil, fmt.Errorf("encountered unrecognised config keys: %v", metadata.Undecoded())
-	}
-
-	return &parsedTeam, nil
-
 }
 
 func fileExists(filename string) bool {
